Allow components to have a requires[] section in their definition.
The array contains a list of components that should be loaded and handled
first prior to loading this one. The component is put on a back-burner
until all requirements are met. Consequently, once the component does get
to render itself, it will call all the required components to create any
needed instances inside the component's just-created DOM tree.